personalização de display e idioma padrão Brasil

+++ SallesWebMvc/SallesWebMvc/Controllers/SellersController.cs
@@ -2,6 +2,8 @@
 using SallesWebMvc.Models;
 using SallesWebMvc.Models.ViewModels;
 using SallesWebMvc.Services;
+using SallesWebMvc.Services.Exceptions;
+using System.Diagnostics;
 
 namespace SallesWebMvc.Controllers
 {
@@ -41,13 +43,13 @@ namespace SallesWebMvc.Controllers
         {
             if(id == null)
             {
-                return NotFound();
+                return RedirectToAction(nameof(Error), new { message = "Id not provided!" });
             }
 
             var obj = _sellerService.FindById(id.Value);
             if(obj == null)
             {
-                return NotFound();
+                return RedirectToAction(nameof(Error), new { message = "Id not found!" });
             }
 
             return View(obj);
@@ -65,29 +67,30 @@ namespace SallesWebMvc.Controllers
         {
             if (id == null)
             {
-                return NotFound();
+                return RedirectToAction(nameof(Error), new { message = "Id not provided!" });
             }
 
             var obj = _sellerService.FindById(id.Value);
             if (obj == null)
             {
-                return NotFound();
+                return RedirectToAction(nameof(Error), new { message = "Id not found!" });
             }
 
             return View(obj);
         }
 
+        //Método GET do Edit
         public IActionResult Edit(int? id)
         {
             if(id == null)
             {
-                return NotFound();
+                return RedirectToAction(nameof(Error), new { message = "Id not provided!" });
             }
 
             var obj = _sellerService.FindById(id.Value);
             if(obj == null)
             {
-                return NotFound();
+                return RedirectToAction(nameof(Error), new { message = "Id not found!" });
             }
 
             List<Department> departments = _departmentService.FindAll();
@@ -100,8 +103,35 @@ namespace SallesWebMvc.Controllers
         [ValidateAntiForgeryToken]
         public IActionResult Edit(int id, Seller seller)
         {
-            _sellerService.Remove(id);
-            return RedirectToAction(nameof(Index));
+            if(id != seller.Id)
+            {
+                return RedirectToAction(nameof(Error), new { message = "Id mismatch!" });
+            }
+
+            try
+            {
+                _sellerService.Remove(id);
+                return RedirectToAction(nameof(Index));
+            }
+            catch (NotFoundException e)
+            {
+                return RedirectToAction(nameof(Error), new { message = e.Message });
+            }
+            catch (DbConcurrencyException e)
+            {
+                return RedirectToAction(nameof(Error), new { message = e.Message });
+            }
+        }
+
+        public IActionResult Error(string message)
+        {
+            var viewModel = new ErrorViewModel
+            {
+                Message = message,
+                RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier
+            };
+
+            return View(viewModel);
         }
     }
 }
